+++ Src/beebwin.h
@@ -255,8 +255,6 @@ public:
 	void CheckForLocalPrefs(const char *path, bool bLoadPrefs);
 	void FindCommandLineFile(char *CmdLineFile);
 	void HandleCommandLineFile(int Drive, const char *CmdLineFile);
-	void HandleEnvironmentVariables();
-	void LoadStartupDisc(int DriveNum, const char *DiscString);
 	bool CheckUserDataPath(bool Persist);
 	void SelectUserDataPath(void);
 	void StoreUserDataPath(void);